+++ tests/testthat/test-plot_methods.R
@@ -416,7 +416,8 @@ test_that("plot_ts correctly applies smoothing and confidence intervals", {
   total_occ_example <- total_occ_ts(example_cube_1,
                                    level = "country",
                                    region = "Denmark")
-
+  total_occ_example <- suppressWarnings(add_ci(total_occ_example,
+                                              num_bootstrap = 10))
   # With confidence intervals
   p2 <- plot_ts(
     total_occ_example,
@@ -514,8 +515,7 @@ test_that("plot_ts handles all parameters without error", {
 
 spec_occ_mammals_denmark_ts <- spec_occ_ts(example_cube_1,
                                            level = "country",
-                                           region = "Denmark",
-                                           ci_type = "none")
+                                           region = "Denmark")
 
 test_that(
   "plot_species_ts returns a ggplot or patchwork object with defaults", {
@@ -577,11 +577,10 @@ test_that("plot_species_ts applies custom aesthetics correctly", {
   expect_equal(point_geom$aes_params$size, 4)
 })
 
-spec_occ_mammals_denmark_ts_ci <- spec_occ_ts(example_cube_1,
+spec_occ_mammals_denmark_ts_ci <- suppressWarnings(spec_occ_ts(example_cube_1,
                                            level = "country",
-                                           region = "Denmark",
-                                           ci_type = "norm",
-                                           num_bootstrap = 20)
+                                           region = "Denmark") %>%
+  add_ci(num_bootstrap = 2))
 
 test_that("plot_species_ts manages trends and confidence intervals", {
   # Verify presence of smoothed trend